Move update mappings to RedisInterface
